Added slideshow to project pages
Added the jQuery files and the css from template one to get the
slideshow to work properly.

Rearranged the layout of the project pages so that the image and main
content is on the left, and the title, date, and short description
(@keywords) is on the left.

Got the fonts to display correctly.