Finishing design and layout
Got the footer to stick to the bottom of the page instead of floating
if there wasn't enough content to push it down.

Also completed the footer layout. had to override the final column
width.

Added patterned backgrounds to the header and footer.

Adjusted the layout on the content pages to match that of the project
pages.

Adjusted the layout of the category pages to match the homepage.
Initially had no idea where the category pages were, come to find out
its really easy to get to them. At the same time figured out how the
naming of the text file in a content folder is how you call a template
file - that was pretty neat.

Added thumbnail images to make the homepage look like nice.

Fixed the spacing of the thumbnails on the homepage, but still have to
go in and fix it on the category pages.