Created full width intro div
Made the intro div on the homepage full width so that it will look
nicer.

Changed the index template so that this would work. Also added the
content to the index text file so that it will pull it from the
index.txt file instead of having to alter the index.html template file.