Fixed the footer listing of recent items.
Changed the recent items list so that it only displays children of the
"first" child of the root.

So basically if creating folders for each year, this will only list the
most recent 5 items in the newest year folder.

To go back to listing the most recent items from each category if not
doing years remove line 3 and line 21 and change the limit on line 7
back to 2.