Added new textures and removed "other" category files
Changed the texture in the intro area and the footer area. The intro
area will most likely be changed again, but for now its good.

Also killed the "Other" category. Only need to have two categories. for
the courses site. Will add more as needed.

Added borders to header, intro, and footer so to create some division
between the sections.

Also added the @description variable to the project pages.