Did so many things!
Fixed the sticky footer across the bottom. Should work in IE now. It
required moving the <hr> push element inside the "wrapper" div on all
template pages.

The projects slider broke for some reason. Tried nivo-gallery first and
that didn't seem to work well. So I changed it to use nivo-slider
instead. It works, but it adds a lot more code and files to the site. l
most likely will try and figure out a way to get the staceyapp included
version to work again.