Got rid of the novo slider
went through and removed the novo slider and replaced it with the
backflip jQuery slider. Used this to make the slideshow touch capable.

Changed the images partial from individual divs for the images to wrap
everything in a list item. This makes it so that the slider works.

Changed the footer background to make it darker.

Changed the header background to make it lighter.

Fixed the footer logo. It now uses the cleaned up M.Davis Courses Logo.
Also added the credits and copyright text.